Normalise both 2FA methods setup
The mobile number controller/form now has been aligned with the
secondary authentication app setup.

The AppAuthForm has been namespaced under App::Authform, following the
same pattern as the Sms::Authform.
