+++ cosmetics-web/app/controllers/secondary_authentication/app_controller.rb
@@ -10,7 +10,7 @@ module SecondaryAuthentication
       user_id = session[:secondary_authentication_user_id]
       return redirect_to(root_path) unless user_id && app_authentication_available?
 
-      @form = AppAuthForm.new(user_id: user_id)
+      @form = App::AuthForm.new(user_id: user_id)
     end
 
     def create
@@ -27,7 +27,7 @@ module SecondaryAuthentication
   private
 
     def form
-      @form ||= AppAuthForm.new(secondary_authentication_params)
+      @form ||= App::AuthForm.new(secondary_authentication_params)
     end
 
     def secondary_authentication_params
