+++ cosmetics-web/app/views/my_account/show.html.erb
@@ -103,7 +103,7 @@
             actions: {
               items: [
                 {
-                  href: edit_my_account_mobile_number_path,
+                  href: new_secondary_authentication_sms_setup_path,
                   text: current_user.sms_authentication_set? ? "Update" : "Add",
                   visuallyHiddenText: "text message"
                 }
