+++ cosmetics-web/config/routes.rb
@@ -24,6 +24,8 @@ Rails.application.routes.draw do
     scope module: "sms", path: "sms" do
       get "not-received", to: "resend#new", as: :new_secondary_authentication_sms_resend
       post "not-received", to: "resend#create", as: :secondary_authentication_sms_resend
+      get "setup", to: "setup#new", as: :new_secondary_authentication_sms_setup
+      post "setup", to: "setup#create", as: :secondary_authentication_sms_setup
     end
 
     get "app", to: "app#new", as: :new_secondary_authentication_app
@@ -163,7 +165,6 @@ Rails.application.routes.draw do
     scope module: :my_account do
       resource :password, controller: :password, only: %i[edit update]
       resource :name, controller: :name, only: %i[edit update]
-      resource :mobile_number, controller: :mobile_number, only: %i[edit update]
       resource :email, controller: :email, only: %i[edit update] do
         member do
           get :confirm
