+++ cosmetics-web/spec/forms/secondary_authentication/sms/setup_form_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe MobileNumberForm do
+RSpec.describe SecondaryAuthentication::Sms::SetupForm do
   subject(:form) do
     described_class.new(user: user, password: password, mobile_number: mobile_number)
   end
